desarrollando mecanica del juego - creando metodo de batalla - version1 - sin API

+++ Clases/Personajes.cs
@@ -16,14 +16,14 @@ namespace EspacioPersonajes
         int fuerza; //1 a 10
         int nivel; //1 a 10
         int defensa; //1 a 10
-        int salud; //100
+        float salud; //100
 
             public int Velocidad { get => velocidad; set => velocidad = value; }
             public int Destreza { get => destreza; set => destreza = value; }
             public int Fuerza { get => fuerza; set => fuerza = value; }
             public int Nivel { get => nivel; set => nivel = value; }
             public int Defensa { get => defensa; set => defensa = value; }
-            public int Salud { get => salud; set => salud = value; }
+            public float Salud { get => salud; set => salud = value; }
     }
 
     public class Datos
